Explain the addition of two sine waves of the same frequency
It would be clearer if the reader knew that adding two sine waves of the same frequency always results in a sine wave of the same frequency but different amplitude and phase.  The using two signals that are 90 degrees out of phase is simply an easier way to generate a particular phase shift.